# Enable rej_uniform native implementation in x86_64