This fixes a bug that caused our rej_uniform implementation to always use the C version instead of correctly using the native implementation when it should.
This bug was introduced, when we added the x86_64 backend as the flag for enabling it was renamed from MLKEM_USE_NATIVE_AARCH64 to MLKEM_USE_NATIVE_REJ_UNIFORM. This was not correctly changed in the rej_uniform.c.

Below are the performance results on my 13th Gen Intel i7-1360P (Raptor Lake) using gcc 14.2.1 from the Arch Linux repo.

Before (6aa6118e):

ML-KEM-512: 22353, 27820, 35663
ML-KEM-768: 39626, 43605, 54916
ML-KEM-1024: 58983, 65402, 80370

After:

ML-KEM-512:  19652, 25165, 32945
ML-KEM-768: 33383, 37343, 48652
ML-KEM-1024: 47658, 54087, 69003

#224