@mkannwischer 2024-11-15
Together with #409, this PR achieves this performance on my machine:
```
ML-KEM-512: 19590, 23861, 30397
ML-KEM-768:  33387, 35734 45130
ML-KEM-1024: 47644, 51798, 64815 
```
That's outperforming the code from the official Kyber repo.