@mkannwischer 2024-11-15
This may be due to some alignment issues. The Kyber code does align the buffer, while we do not. 
I don't have time to investigate now. I'll look at it later.