@hanno-becker 2024-11-16
@mkannwischer We can't have undocumented padding assumptions -- we either need to complicate our native API here, or fix this in a different way.

I tried changing the AVX2 code to not overread, and removed the AVX 128-bit tail. The performance numbers before/after are here:

```
Measured on c7i.xlarge

Without 128-bit tail:

   keypair cycles = 22463
    encaps cycles = 29165
    decaps cycles = 38154

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  22186  22274  22331  22366  22403  22463  22549  22805  22987  23263  24488
    encaps percentiles:  28850  28950  29005  29054  29102  29165  29247  29386  29595  30027  31129
    decaps percentiles:  37851  37958  38018  38055  38100  38154  38250  38417  38642  38936  40530

    keypair cycles = 38740
    encaps cycles = 42251
    decaps cycles = 55424

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  38027  38278  38421  38549  38634  38740  38864  39007  39252  39813  41013
    encaps percentiles:  41790  41937  42038  42111  42185  42251  42356  42480  42690  43198  44812
    decaps percentiles:  54762  54974  55090  55212  55294  55424  55554  55718  55926  56517  57886

    keypair cycles = 53546
    encaps cycles = 59473
    decaps cycles = 77068

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  52816  53152  53248  53352  53453  53546  53623  53746  53919  55186  56879
    encaps percentiles:  58921  59120  59223  59312  59377  59473  59592  59721  59982  61101  62263
    decaps percentiles:  76332  76636  76764  76878  76972  77068  77191  77341  77592  78606  80070

====================

With 128-bit tail:

    keypair cycles = 22046
    encaps cycles = 28421
    decaps cycles = 37342

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21872  21930  21955  21977  22001  22046  22218  22421  22631  23021  24730
    encaps percentiles:  28252  28331  28360  28383  28400  28421  28446  28491  28661  29045  30517
    decaps percentiles:  37144  37204  37246  37286  37311  37342  37413  37522  37746  38312  39502

    keypair cycles = 37742
    encaps cycles = 41222
    decaps cycles = 54438

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  37065  37232  37378  37494  37615  37742  37902  38186  38579  39220  40190
    encaps percentiles:  41046  41107  41138  41162  41190  41222  41302  41525  41631  42267  43848
    decaps percentiles:  53825  53927  54006  54088  54266  54438  54583  54783  55070  55799  57065

    keypair cycles = 52383
    encaps cycles = 57990
    decaps cycles = 75675

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  51865  52079  52182  52248  52312  52383  52465  52540  52690  54124  56110
    encaps percentiles:  57702  57809  57851  57892  57939  57990  58073  58174  58368  59857  60602
    decaps percentiles:  74983  75234  75360  75478  75582  75675  75801  75953  76295  77272  78286
```

We loose a little bit of performance by not doing the 128-bit tail, so I'll have another look if there's a simple way to fix it. But if not, I think this is still a more robust approach than going via padding.