@hanno-becker 2024-11-16
With shortened 128-bit tail:
```
    keypair cycles = 22046
    encaps cycles = 28440
    decaps cycles = 37288

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21874  21938  21961  21985  22008  22046  22163  22412  22623  23001  24065
    encaps percentiles:  28328  28366  28388  28406  28423  28440  28464  28509  28646  29177  30683
    decaps percentiles:  37116  37169  37199  37235  37261  37288  37331  37413  37576  38241  39548

    keypair cycles = 37763
    encaps cycles = 41192
    decaps cycles = 54268

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  37112  37252  37424  37552  37653  37763  37900  38129  38438  39192  40556
    encaps percentiles:  41027  41069  41096  41126  41153  41192  41277  41470  41594  42163  43596
    decaps percentiles:  53790  53875  53948  54005  54129  54268  54393  54606  54863  55519  56650

    keypair cycles = 52031
    encaps cycles = 58043
    decaps cycles = 75697

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  51544  51701  51828  51906  51961  52031  52106  52221  52329  53832  55546
    encaps percentiles:  57752  57845  57886  57930  57987  58043  58174  58317  58547  59925  61176
    decaps percentiles:  74945  75227  75359  75477  75544  75697  75860  76065  76412  77446  79046
```